Refactor protocol utils
